fix route and another UI

+++ src/components/plan/InComplete.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useContext } from "react";
 import { GlobalContext } from "context/GlobalState";
-
+import Swal from "sweetalert2";
 import { formatRp } from "utils/formatRp";
 
 const InComplete = () => {
@@ -22,8 +22,20 @@ const InComplete = () => {
   console.log(totalPrice);
 
   const handleDeletePlan = (planId) => {
-    console.log(planId, "ini planId");
-    deletePlan(planId);
+    Swal.fire({
+      title: "Hapus Plan?",
+      text: "Kamu tidak bisa membatalkan ini!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deletePlan(planId);
+        Swal.fire("Deleted!", "Your file has been deleted.", "success");
+      }
+    });
   };
 
   return (
